PlanConfig.Valor de int a decimal(18,2): soportar precios con centavos
- BD Ferozo: ALTER TABLE PlanConfig ALTER COLUMN Valor decimal(18,2)
- Entidad PlanConfig: int -> decimal con Column annotation
- DTO PlanParametro: int -> decimal, formateo inteligente (N0 si entero, N2 si decimal)
- PlanComparacion helpers: cast explicito a int para Max*, decimal para Importe*
- LicenciaService: Convert.ToDecimal en mapeo desde Dapper
- RepositorioLicencias: todos los metodos adaptados (GuardarPlanConfigAsync, AgregarPlanConfigAsync, GetPlanValoresAsync)
- UpgradePago.razor: Dictionary<string,decimal>, FormatValor decimal
- PlanConfiguracion.razor: input step=0.01, editValor decimal, display N2 para decimales
- Program.cs: casts explicitos en activacion Free
- Precios muestran USD prefix en Portal MT

+++ PortalLicencias/DigitalPlus.Licencias/Program.cs
@@ -308,10 +308,10 @@ app.MapPost("/api/activar-free", async (ActivarFreeRequest req,
 
     // Obtener limites del plan Free de PlanConfig
     var planValores = await repo.GetPlanValoresAsync("free");
-    int maxLegajos = planValores.GetValueOrDefault("MaxLegajos", 5);
-    int maxSucursales = planValores.GetValueOrDefault("MaxSucursales", 1);
-    int maxFichadasMes = planValores.GetValueOrDefault("MaxFichadasRolling30d", 200);
-    int duracionDias = planValores.GetValueOrDefault("DuracionDias", 0);
+    int maxLegajos = (int)planValores.GetValueOrDefault("MaxLegajos", 5);
+    int maxSucursales = (int)planValores.GetValueOrDefault("MaxSucursales", 1);
+    int maxFichadasMes = (int)planValores.GetValueOrDefault("MaxFichadasRolling30d", 200);
+    int duracionDias = (int)planValores.GetValueOrDefault("DuracionDias", 0);
 
     var dbName = config["MultiTenant:DatabaseName"] ?? "DigitalPlusMultiTenant";
     DigitalPlus.Licencias.Entidades.Empresa? empresa = null;
